对可以正常运行代码的进一步修改
# DETAILS:
- API 的重命名
- 函数类型去掉 `np.float64`, 默认使用的就是这个所以去掉, 让代码一行的长度能减少一些

+++ opencood/data_utils/data_models/dataset_models.py
@@ -39,7 +39,7 @@ class CAVData(BaseModel):
     camera_data: SkipValidation[List[Image.Image]]
     depth_data: SkipValidation[List[Image.Image]]
 
-    lidar_np: Optional[np.ndarray[np.float32]] = None
+    lidar_np: Optional[np.ndarray[np.float64]] = None
     modality_name: Optional[str] = None
     file_extensions: Dict = Field(default_factory=dict)
